평가정보 수정

+++ src/main/java/com/namecard/indicator/dto/result/IndicatorInfoResult.java
@@ -20,6 +20,6 @@ public class IndicatorInfoResult {
     @Builder
     public static class IndicatorScoreResult {
         private String tagName;
-        private int tagScore;
+        private double tagScore;
     }
 }
